+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 - Add `GET` support in Apple II Basic. Add relevant unit tests.
 - Add new `input` package to properly handle inputs in Ebiten, TTY and unit test modes.
 - Add support for real and integer variables in GET instruction
+- Distribute example files and programs into subdirectories.
 
 ### Changed
 - Refactor input/output devices implementation.